+++ test.json
@@ -1,13 +1,13 @@
 {
-    "wow": {
+    "ow": {
         "type": "INPUT_NODE", 
         "hierarchy": 0
-    }, 
-    "kami": {
+        }, 
+    "karsa": {
         "type": "INPUT_NODE", 
         "hierarchy": 0
-    }, 
-    "denki": {
+        }, 
+    "wen": {
         "type": "OUTPUT_NODE", 
         "hierarchy": 3, 
         "inputs": ["and_1"]
@@ -15,20 +15,20 @@
     "and_1": {
         "type": "BLOCK", 
         "block_type": "AND", 
-        "inputs": ["wow", "kami"], 
+        "inputs": ["ow", "karsa"], 
         "output": "and_1", 
         "hierarchy": 1
     }, 
     "wire_1": {
         "type": "WIRE", 
         "from": "and_1", 
-        "to": "denki", 
+        "to": "wen", 
         "hierarchy": 2
     }, 
     "wire_2": {
         "type": "WIRE", 
         "from": "and_1", 
-        "to": "denki", 
+        "to": "wen", 
         "hierarchy": 2
     }
 }
